+++ gnovm/cmd/gno/testdata/gno_transpile/01_no_args.txtar
