+++ gnovm/cmd/gno/testdata/gno_transpile/02_empty_dir.txtar
