+++ gnovm/cmd/gno/testdata/gno_transpile/03_gno_files_parse_error.txtar
