+++ gnovm/cmd/gno/testdata/gno_transpile/04_valid_gno_files.txtar
