+++ gnovm/cmd/gno/testdata/gno_transpile/05_skip_fmt_flag.txtar
