+++ gnovm/cmd/gno/testdata/gno_transpile/06_build_flag.txtar
