+++ gnovm/cmd/gno/testdata/gno_transpile/09_gno_files_whitelist_error.txtar
